Сделано распределение по подмодулям файла с базой данных, для более красивого чтения и понимания кода программы, так же обновлены докумнетация на неё, расставлены заголовки для понимания, что за участок кода, а так же сделан импорт и экспорт публичного API для работы с этими функциями
